Complete ExcelBuilder and refactor all codes based upon new builder design pattern
