+++ ExcelWizard/Models/ExcelBuilderContracts.cs
@@ -11,9 +11,13 @@ public interface IExcelBuilder
 public interface IExpectGeneratingExcelTypeExcelBuilder
 {
     /// <summary>
+    /// Generate simple Grid layout Excel file with the option of simply using a model bind and configure Excel with attributes
+    /// </summary>
+    IExpectGridLayoutExcelBuilder CreateGridLayoutExcel();
+
+    /// <summary>
     /// The Excel is not a grid layout Excel, therefore cannot be created through binding to a model and would be created composing different sub-components and configs
     /// </summary>
-    /// <returns></returns>
     IExpectSheetsExcelBuilder CreateComplexLayoutExcel();
 }
 
@@ -45,10 +49,35 @@ public interface IExpectStyleExcelBuilder
 
 public interface IExpectOtherPropsAndBuildExcelBuilder : IExpectBuildExcelBuilder
 {
+    /// <summary>
+    /// Set the default IsLocked value for all Sheets. Default is false
+    /// </summary>
     IExpectBuildExcelBuilder SetDefaultLockedStatus(bool isLockedByDefault);
 }
 
 public interface IExpectBuildExcelBuilder
 {
     ExcelModel Build();
+}
+
+public interface IExpectGridLayoutExcelBuilder
+{
+    /// <summary>
+    /// Generate Simple Single Sheet Grid layout Excel file from special model configured options with [ExcelSheet] and [ExcelSheetColumn] attributes
+    /// </summary>
+    /// <param name="bindingDataListModel"> List of data (should be something like List{Person}()) </param>
+    IExpectBuildExcelBuilder WithOneSheetUsingAModelToBind(object bindingDataListModel);
+
+    /// <summary>
+    /// Generate Grid layout Excel having multiple Sheets from special model configured options with [ExcelSheet] and [ExcelSheetColumn] attributes
+    /// </summary>
+    /// <param name="listOfBindingDataListModel"></param>
+    /// <returns> List of data list. e.g. object list of Persons, Phones, Universities, etc which each will be mapped to a Sheet </returns>
+    IExpectStyleExcelBuilder WithMultipleSheetsUsingModelListToBind(List<object> listOfBindingDataListModel);
+
+    /// <summary>
+    /// Generate Grid layout Excel in usual way by create Sheets manually step by step and without model binding
+    /// </summary>
+    /// <returns></returns>
+    IExpectSheetsExcelBuilder ManuallyWithoutModelBinding();
 }
